chore: extract A11y helpers
